+++ Assets/3_Scripts/Data/DataClasses/Stat/BaseStat.cs
@@ -7,9 +7,9 @@ public class BaseStat
     // 행위자의 식별 정보 : ID는 몬스터와 플레이어를 나눠서 따로 저장할 생각
     [SerializeField] protected string actorName;
     // 행위자의 공통 변수
-    [SerializeField] protected float maxHp;
+    [SerializeField] protected int maxHp;
     [SerializeField] protected float speed;
-    [SerializeField] protected float defenceValue;
+    [SerializeField] protected int defenceValue;
     #endregion
 
     #region Public : Property
@@ -17,8 +17,8 @@ public class BaseStat
     public string GetActorName { get { return actorName; } }
 
     // 행위자의 공통 변수
-    public float MaxHP { get { return maxHp; } set { maxHp = value; } }
+    public int MaxHP { get { return maxHp; } set { maxHp = value; } }
     public float Speed { get { return speed; } set { speed = value; } }
-    public float Defence { get { return defenceValue; } set { defenceValue = value; } }
+    public int Defence { get { return defenceValue; } set { defenceValue = value; } }
     #endregion
 }
